// src/lib/counterparty/api.d.ts
export interface AssetInfo {
    asset_longname: string | null
    description: string
    issuer: string | null
    divisible: boolean
    locked: boolean
}

export interface XCPAPIDispense {
    asset: string
    asset_info: AssetInfo
    block_index: number
    block_time: number
    btc_amount: number
    btc_amount_normalized: string
    destination: string
    dispense_index: number
    dispense_quantity: number
    dispense_quantity_normalized: string
    dispenser: XCPAPIDispenser
    dispenser_tx_hash: string
    source: string
    tx_hash: string
    tx_index: number
}

export type DispenserStatus = "open" | "closed" | "closing" | "open_empty_address"

export interface XCPAPIDispenser {
    tx_index: number
    tx_hash: string
    block_index: number
    source: string
    asset: string
    give_quantity: number
    escrow_quantity: number
    satoshirate: number
    status: DispenserStatus
    give_remaining: number
    oracle_address: string | null
    last_status_tx_hash: string | null
    origin: string
    dispense_count: number
    last_status_tx_source: string | null
    close_block_index: number | null
    price: number
    fiat_price: number | null
    oracle_price: number | null
    fiat_unit: number | null
    oracle_price_last_updated: number | null
    satoshi_price: number
    block_time: number
    asset_info?: AssetInfo
    give_quantity_normalized: string
    give_remaining_normalized: string
    escrow_quantity_normalized: string
    satoshirate_normalized: string
    satoshi_price_normalized: string
    price_normalized: string
}